♻️Refactor: x 아이콘 통일
